parallelization and file mvmnt changes
instead of moving all files at the end, each invocation of rMutant also has a `mv` command for specific files based on the arguments that get passed to rMutant, so each process can take care of its own file operations. This specificity introduces a couple of failure points for mv, such as when rMutant is told to mutate a residue from Q to Q, for example, and it doesn't output anything, so mv has nothing to move. I just pipe the output to /dev/null, errors and all, which is slightly dirty, but I think not enough to warrant a more sophisticated but convoluted fix.